get rid of some unnecessary complexity
1. a global semidet_sccs_cache variable was storing a Boolean vector
(for the bscc_avoid optimization) for each processed automaton,
without ever releasing the memory, and this map was queried for each
state/edge/scc multiple time during the construction.  It seems easier
to simply store this vector as a member of the bp_twa class, skipping
the need for the global cache and associated lookups.

2. the cut_condition function, which also uses the above Boolean
vector, was declared outside of the bp_twa class and passed as an
argument.  However it's the only value ever used for this argument, so
let's simplify everything and make it a member function, with access
to the said vector.